+++ server/src/api/socket/asset_manager/ddraft.py
@@ -36,12 +36,6 @@ def get_ddraft_data(data: bytes):
 
     image = base64.b64decode(ddraft_file["image"])
 
-    template = {
-        "ddraft/uvtt": {
-            "width": ddraft_file["resolution"]["map_size"]["x"] * 50,
-            "height": ddraft_file["resolution"]["map_size"]["y"] * 50,
-            "options": json.dumps([[f"ddraft_{k}", v] for k, v in ddraft_file.items() if k != "image"]),
-        }
-    }
+    template = { k: v for k, v in ddraft_file.items() if k != "image" }
 
     return image, template