+++ server/src/save.py
@@ -15,7 +15,7 @@ When writing migrations make sure that these things are respected:
 - It's often a good idea to start the server with a clean save and use `.schema <table_name>` in sqlite to get the exact schema output that a clean save creates
 """
 
-SAVE_VERSION = 118
+SAVE_VERSION = 119
 
 import asyncio
 import json
@@ -895,6 +895,10 @@ def upgrade(
                 "UPDATE location_options SET ambient_light = NULL WHERE id NOT IN (SELECT default_options_id FROM room)"
             )
             db.execute_sql('ALTER TABLE "aura" ADD COLUMN "flood_light" INTEGER NOT NULL DEFAULT 0')
+    elif version == 118:
+        # Add Asset.kind_specific_data
+        with db.atomic():
+            db.execute_sql('ALTER TABLE "asset" ADD COLUMN "kind_specific_data" BLOB DEFAULT NULL')
     else:
         raise UnknownVersionException(f"No upgrade code for save format {version} was found.")
     inc_save_version(db)